[Update] combine ProjectModalWriting with ProjectModalEdit
